Membership protocol implementation initial commit
This MP (Machine Programming assignment) is about implementing a
membership protocol similar to one we discussed in class. Since
it is infeasible to run a thousand cluster nodes (peers) over a
real network, we are providing you with an implementation of an
emulated network layer (EmulNet). Your membership protocol
implementation will sit above EmulNet in a peer- to-peer (P2P)
layer, but below an App layer. Think of this like a three layer
protocol stack with Application, P2P, and EmulNet as the three
layers (from top to bottom).